pyproject.toml
[project]
name = "yolos-face-detection"
version = "0.1.0"
description = "YOLOS for face detection"
license = { text = "Apache-2.0" }
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "albumentations>=2.0.8",
    "pillow>=12.0.0",
    "rich>=14.2.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
    "opencv-python>=4.12",
]


[build-system]
requires = ["uv_build>=0.9.13,<0.10.0"]
build-backend = "uv_build"


[tool.uv.build-backend]
module-name = "yolos_face"

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "notebook>=7.5.0",
]
